refactor:핀하우스 공고리스트 / 공고 필터 / 역할 책임 분리
